register squashed-normal-type. pass test for drq
