+++ torchbenchmark/models/soft_actor_critic/nets.py
@@ -5,7 +5,7 @@ from torch import distributions as pyd
 from torch import nn
 
 from . import utils
-
+from torchbenchmark.util.distribution import SquashedNormal
 
 def weight_init(m):
     if isinstance(m, nn.Linear):
@@ -119,7 +119,9 @@ class StochasticActor(nn.Module):
                 self.log_std_high - self.log_std_low
             ) * (log_std + 1)
             std = log_std.exp()
-            dist = SquashedNormal(mu, std)
+            tanh_transform_clamp_min = -0.99
+            tanh_transform_clamp_max = 0.99
+            dist = SquashedNormal(mu, std, tanh_transform_clamp_min, tanh_transform_clamp_max)
         elif self.dist_impl == "beta":
             out = 1.0 + F.softplus(out)
             alpha, beta = out.chunk(2, dim=1)
@@ -201,54 +203,6 @@ class BetaDist(pyd.transformed_distribution.TransformedDistribution):
     @property
     def mean(self):
         mu = self.base_dist.mean
-        for tr in self.transforms:
-            mu = tr(mu)
-        return mu
-
-
-"""
-Credit for actor distribution code: https://github.com/denisyarats/pytorch_sac/blob/master/agent/actor.py
-"""
-
-
-class TanhTransform(pyd.transforms.Transform):
-    domain = pyd.constraints.real
-    codomain = pyd.constraints.interval(-1.0, 1.0)
-    bijective = True
-    sign = +1
-
-    def __init__(self, cache_size=1):
-        super().__init__(cache_size=cache_size)
-
-    @staticmethod
-    def atanh(x):
-        return 0.5 * (x.log1p() - (-x).log1p())
-
-    def __eq__(self, other):
-        return isinstance(other, TanhTransform)
-
-    def _call(self, x):
-        return x.tanh()
-
-    def _inverse(self, y):
-        return self.atanh(y.clamp(-0.99, 0.99))
-
-    def log_abs_det_jacobian(self, x, y):
-        return 2.0 * (math.log(2.0) - x - F.softplus(-2.0 * x))
-
-
-class SquashedNormal(pyd.transformed_distribution.TransformedDistribution):
-    def __init__(self, loc, scale):
-        self.loc = loc
-        self.scale = scale
-
-        self.base_dist = pyd.Normal(loc, scale)
-        transforms = [TanhTransform()]
-        super().__init__(self.base_dist, transforms)
-
-    @property
-    def mean(self):
-        mu = self.loc
         for tr in self.transforms:
             mu = tr(mu)
         return mu
